Fix FS caching issues in the metastore
The Warehouse has a single configuration object. The configuration is
passed down to the FS when doing FS operations. Unfortunately, the
HopsSSLSocket factory changes the configuration and adds client
keystore/truststore information.

The keystore information is also used to get cached FS objects from the
cache. As the "root configuration" object is shared across Metastore clients
and users, this makes that in case of concurrent operations users might
get from the cache a FS object for another user. Triggering either an
authentication exception (UGI doesn't match the certificates) or a
Certificates not found exception as they might have been cleaned up upon
the right user closing the connection.

This is a temporary workaround only limited to the metastore. I believe
a proper solution is to have the HopsSSLSocket factory to clone the
configuration object before injecting user specific configurations.