Validate ssl-server.xml permissions before loading
Hive Server and Hive Metastore rely on HiveConf object. To setup the
thrift server correctly they also need access to the ssl-server.xml.

The same configuration object is also used by the clients, which don't
have access to that file. There is already a flag to disable the
loading, however in some cases it's not possible to distinguish if the
call originates from a server or a client. As such, we add a validation
step that checks the permissions of the file before loading. If the file
cannot be read than it's skipped.

Moreover add a the JWT token properties in the ssl-server file in
the hidden list for both the Hive Server and Hive Metastore